requirements.txt
asgiref==3.3.1
awesome-slugify==1.6.5
beautifulsoup4==4.6.3
bleach==2.1.4
certifi==2018.8.24
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1
django-appconf==1.0.2
django-cors-headers==2.4.0
django-crispy-forms==1.7.2
django-emoticons==1.2
django-filter==2.0.0
django-widget-tweaks==1.4.3
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
easy-thumbnails==2.5
html5lib==1.0.1
idna==2.7
isort==4.3.21
Markdown==3.0.1
olefile==0.46
Pillow==8.1.2
pyflakes==2.1.1
PyJWT==1.6.4
python-decouple==3.1
pytz==2018.5
regex==2021.3.17
requests==2.19.1
six==1.11.0
sorl-thumbnail==12.5.0
sqlparse==0.4.1
Unidecode==0.4.21
urllib3==1.23
webencodings==0.5.1
yapf==0.30.0
